refactor(tools): moved out & split out all propertyGroups related code  * renamed process_properties & process_prefixItems to process_structs & process_tupples respectively  * a bit of extra cleanup
