+++ tools/bevy_components/propGroups/process_component.py
@@ -1,0 +1,89 @@
+import bpy
+from bpy_types import PropertyGroup
+from bpy.props import (PointerProperty)
+from . import process_structs
+from . import process_tupples
+from . import process_enum
+from . import process_list
+
+def process_component(registry, definition, update, extras=None, component_name_override=None):
+    component_name = definition['title']
+    short_name = definition["short_name"]
+    type_info = definition["typeInfo"] if "typeInfo" in definition else None
+    type_def = definition["type"] if "type" in definition else None
+    properties = definition["properties"] if "properties" in definition else {}
+    prefixItems = definition["prefixItems"] if "prefixItems" in definition else []
+
+    has_properties = len(properties.keys()) > 0
+    has_prefixItems = len(prefixItems) > 0
+    is_enum = type_info == "Enum"
+    is_list = type_info == "List"
+
+    print("processing", short_name, "component_name_override", component_name_override)
+
+    __annotations__ = {}
+    tupple_or_struct = None
+
+    with_properties = False
+    with_items = False
+    with_enum = False
+    with_list = False
+
+    print("entry", component_name, type_def, type_info)# definition)
+
+    if has_properties:
+        __annotations__ = __annotations__ | process_structs.process_structs(registry, definition, properties, update, component_name_override)
+        with_properties = True
+        tupple_or_struct = "struct"
+
+    if has_prefixItems:
+        __annotations__ = __annotations__ | process_tupples.process_tupples(registry, definition, prefixItems, update, None, component_name_override)
+        with_items = True
+        tupple_or_struct = "tupple"
+
+    if is_enum:
+        __annotations__ = __annotations__ | process_enum.process_enum(registry, definition, update, component_name_override)
+        with_enum = True
+
+    if is_list:
+        __annotations__ = __annotations__ | process_list.process_list(registry, definition, update, component_name_override)
+        with_list= True
+    
+    field_names = []
+    for a in __annotations__:
+        field_names.append(a)
+    single_item = len(field_names)
+
+    extras = extras if extras is not None else {
+        "type_name": component_name
+    }
+
+    print("DONE: __annotations__", __annotations__)
+    print("")
+    property_group_name = short_name+"_ui"
+    property_group_params = {
+         **extras,
+        '__annotations__': __annotations__,
+        'tupple_or_struct': tupple_or_struct,
+        'single_item': single_item, 
+        'field_names': field_names, 
+        **dict(with_properties = with_properties, with_items= with_items, with_enum= with_enum, with_list= with_list, short_name= short_name),
+        #**dict(update = update_test)
+    }
+    (property_group_pointer, property_group_class) = property_group_from_infos(property_group_name, property_group_params)
+    # add our component propertyGroup to the registry
+    registry.register_component_propertyGroup(property_group_name, property_group_pointer)
+    # for practicality, we add an entry for a reverse lookup (short => long name, since we already have long_name => short_name with the keys of the raw registry)
+    registry.add_shortName_to_longName(short_name, component_name)
+
+    return (property_group_pointer, property_group_class)
+                
+def property_group_from_infos(property_group_name, property_group_parameters):
+    # print("creating property group", property_group_name)
+    property_group_class = type(property_group_name, (PropertyGroup,), property_group_parameters)
+    
+    bpy.utils.register_class(property_group_class)
+    property_group_pointer = PointerProperty(type=property_group_class)
+    #setattr(bpy.types.Object, property_group_name, property_group_pointer)
+    
+    return (property_group_pointer, property_group_class)
