+++ tools/bevy_components/propGroups/utils.py
@@ -1,0 +1,3 @@
+# helper function that returns a lambda, used for the PropertyGroups update function
+def update_calback_helper(definition, update, component_name_override):
+    return lambda self, context: update(self, context, definition, component_name_override)
